[CodingStandard] decopule ClassRefactor from ClassWrapper
